feat: Adding device images
So far I've only gone though a fraction of the devices to collect the images
and I need to work out a better naming and attribution tracking stratagy.

The images are currently only visible within the expanded panel, but the first
image in the list will be added to the list itself at some point.

Also bumped some version numbers whilst I was at it.